+++ CHANGELOG.md
@@ -26,6 +26,7 @@ No ticket - change admin confirm url
 - TT-2187 - Add django admin filters for GDPR, publish place and verification method
 - TT-2198 - TestApi: add endpoint to delete companies created by automated tests
 - TT-1304 - Upgrade sentry client
+- TT-2216 - TestApi: add endpoint to delete buyers created by automated tests
 
 ### Fixed bugs
 - TT-2064 - Submit user details with request to verify